requirement.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Deprecated==1.2.14
Django==4.2.3
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.11
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
vonage==3.7.1
vonage-jwt==1.0.0
wrapt==1.15.0
